@@ -14,28 +14,31 @@
         "write-heading-ids": "docusaurus write-heading-ids"
     },
     "dependencies": {
-        "@docusaurus/core": "^2.4.1",
-        "@docusaurus/preset-classic": "^2.4.1",
-        "@docusaurus/theme-mermaid": "^2.4.1",
-        "@mdx-js/react": "^1.6.22",
-        "clsx": "^1.2.0",
-        "lucide-react": "^0.216.0",
-        "prism-react-renderer": "^1.3.5",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "redocusaurus": "^1.6.2"
+        "@docusaurus/core": "^3.0.0",
+        "@docusaurus/preset-classic": "^3.0.0",
+        "@docusaurus/theme-mermaid": "^3.0.0",
+        "@mdx-js/react": "^3.0.0",
+        "clsx": "^2.0.0",
+        "lucide-react": "^0.292.0",
+        "prism-react-renderer": "^2.1.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "redocusaurus": "^2.0.0"
     },
     "devDependencies": {
-        "@docusaurus/module-type-aliases": "^2.4.1",
-        "@tsconfig/docusaurus": "^1.0.7",
-        "@typescript-eslint/eslint-plugin": "^5.59.2",
-        "@typescript-eslint/parser": "^5.59.2",
-        "eslint": "^8.37.0",
-        "eslint-plugin-react": "^7.32.2",
+        "@docusaurus/module-type-aliases": "^3.0.0",
+        "@tsconfig/docusaurus": "^2.0.2",
+        "@typescript-eslint/eslint-plugin": "^6.9.1",
+        "@typescript-eslint/parser": "^6.9.1",
+        "eslint": "^8.53.0",
+        "eslint-plugin-react": "^7.33.2",
         "eslint-plugin-require-explicit-generics": "^0.4.2",
         "fork-ts-checker-webpack-plugin": "^6.5.3",
         "prettier": "2.8.4",
-        "typescript": "^5.0.4"
+        "typescript": "^5.2.2",
+        "@docusaurus/types": "3.0.0",
+        "@docusaurus/tsconfig": "^3.0.0",
+        "@types/react": "^18.2.34"
     },
     "browserslist": {
         "production": [
@@ -50,7 +53,7 @@
         ]
     },
     "engines": {
-        "node": ">=16.14"
+        "node": ">=18.0"
     },
     "overrides": {
         "trim": "^0.0.3",
